requirements.txt
azure-functions
azure-iot-device